(T): Added test for backtrace source code links
